Refactor location of collect_checkpoint_paths
* Moved from tacotron/evaluate.py to tacotron/input/helpers.py
